+++ src/components/rootPage/Posters/index.js
@@ -1,0 +1,2 @@
+export { default as Posters } from './Posters';
+export { default as Poster } from './Poster';
